--- smsapiTests/Contacts/CreateGroupTest.cs
@@ -1,30 +1,0 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using SMSApi.Api;
-using SMSApi.Api.Response;
-
-namespace smsapiTests.Contacts
-{
-    [TestClass]
-    public class CreateGroupTest : ContactsTestBase
-    {
-        [TestInitialize]
-        [TestCleanup]
-        public void Cleanup()
-        {
-            var groups = _factory.ListGroups().Execute();
-            foreach (var group in groups.Collection)
-            {
-                _factory.DeleteGroup(group.Id).Execute();
-            }
-        }
-
-        [TestMethod]
-        public void TestCreateGroup()
-        {
-            var group = _factory.CreateGroup().SetName("exampleGroup").Execute();
-
-            Assert.AreEqual("exampleGroup", group.Name);
-            Assert.IsNotNull(group.Id);
-        }
-    }
-}
